nonport.cpp: applyToDirContents: Fix bug: 'handle' must be 'intptr_t'.
This bug is very old, but was only uncovered by running with MSYS2
mingw-w64, which evidently uses the 64-bit _findfile, and hence
requires the proper type be used to hold the result.

+++ nonport.cpp
@@ -336,7 +336,7 @@ void applyToDirContents(char const *dirName,
     strcpy(buf, dirName);
     if (buf[strlen(buf)-1] != '\\') strcat(buf, "\\");
     strcat(buf, "*");
-    long handle = _findfirst(buf, &fb);
+    intptr_t handle = _findfirst(buf, &fb);
     delete buf;
     int done = (handle == -1);
     if (handle == -1 && errno != ENOENT) // ENOENT = no matching entries
